rewrite fitness tests as browser tests
